add custom dataloader for ssl
