+++ config.txt
@@ -2,22 +2,22 @@ seed = 0
 num_workers = 6
 adam = 1
 initialization = 0
-arch = resnet50
+arch = resnet50_scratch
 target_val_accuracy = 95.0
 test_model = True
 model_in_name = resnet50-cifar10.pt
 model_out_name = resnet50-cifar10.pt
-model_weights_dir = /work/netthinker/shared/ayush/model_weights/
-p_batch_size = 128
+model_weights_dir = ./model_weights/
+p_batch_size = 8
 p_lr = 0.004
 p_momentum = 0.9
 p_weight_decay = 0.0
 p_num_epochs = 5
-t_batch_size = 128
+t_batch_size = 8
 t_lr = 0.004
 t_momentum = 0.9
 t_weight_decay = 0.0
 t_num_epochs = 100
 experiment_name = <comet_exp_name>
-dataset = /work/netthinker/shared/cifar10
-
+dataset = ../../cifar10
+gpu_avail = 0
