+++ resnet.py
@@ -1,3 +1,4 @@
+import numpy as np
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
@@ -98,20 +99,38 @@ class ResNet(nn.Module):
         self.avgpool = nn.AdaptiveAvgPool2d((1, 1))
         self.fc = nn.Linear(512*ResBlock.expansion, num_classes)
 
-    def forward(self, x):
-        x = self.relu(self.batch_norm1(self.conv1(x)))
-        x = self.max_pool(x)
+    def get_grid_col(self, grid_tensor):
+        arr = grid_tensor.detach().cpu().numpy()
 
-        x = self.layer1(x)
-        x = self.layer2(x)
-        x = self.layer3(x)
-        x = self.layer4(x)
+    def forward(self, y):
 
-        x = self.avgpool(x)
-        x = x.reshape(x.shape[0], -1)
-        x = self.fc(x)
+        new_x = torch.transpose(y, 0, 1)
+        # print(f"Here: {new_x.shape}")
+        outputs = []
+        for x in new_x:
+            # print(x.shape)
+            x = self.relu(self.batch_norm1(self.conv1(x)))
+            x = self.max_pool(x)
 
-        return x
+            x = self.layer1(x)
+            x = self.layer2(x)
+            x = self.layer3(x)
+            x = self.layer4(x)
+
+            x = self.avgpool(x)
+            x = x.reshape(x.shape[0], -1)
+            x = self.fc(x)
+            # (8,4)
+            # print(x.shape, type(x))
+            outputs.append(x)
+
+        # (4,8,4)
+        # 8,4,4
+        outputs = torch.stack(outputs)
+        outputs = torch.transpose(outputs, 0, 1)
+
+        # print(f"Outputs: {outputs.shape}, {type(outputs)}")
+        return outputs
 
     def _make_layer(self, ResBlock, blocks, planes, stride=1):
         ii_downsample = None
